pyproject.toml
[tool.poetry]
name = "dejure"
version = "0.1.0"
description = ""
authors = ["Nickolsky <romb.intu@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pymongo = "^4.5.0"
aiogram = "^3.1.1"
python-dotenv = "^1.0.0"
loguru = "^0.7.2"


[tool.poetry.group.dev.dependencies]
pymongo-inmemory = "^0.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
